+++ src/utils/constant.js
@@ -1,0 +1,6 @@
+import R from 'ramda'
+
+export const NO_BREAK_SPACE = 160
+
+export const PRINTABLE_CHAR_CODES = R.range(33, 127)
+export const NO_BREAK_SPACE_CHAR_CODE = NO_BREAK_SPACE