+++ src/utils/scramblers.js
@@ -1,38 +1,39 @@
 import R from 'ramda'
 
-const START_CODE = 33
-const END_CODE = 126
-const RANGE = END_CODE - START_CODE
+import {
+  PRINTABLE_CHAR_CODES,
+  NO_BREAK_SPACE_CHAR_CODE,
+} from './constant'
 
-const randomChar = R.pipe(
-  Math.random,
-  R.multiply(RANGE),
-  Math.floor,
-  R.add(START_CODE),
+const noBreakSpace = String.fromCharCode(NO_BREAK_SPACE_CHAR_CODE)
+
+const randomChar = () => R.pipe(
+  R.concat([NO_BREAK_SPACE_CHAR_CODE]),
+  array => array[Math.floor(Math.random() * array.length)],
   String.fromCharCode,
-)
+)(PRINTABLE_CHAR_CODES)
 
 const array2String = R.reduce(R.concat, '')
 
-export const mixcramble = (_, text = "", mask = []) => R.pipe(
+export const mixcramble = (_, text = "", mask = [], noBreakSpaceFlag) => R.pipe(
   R.addIndex(R.map)((d, i) => d === 0
     ? randomChar()
-    : text[i] || ''
+    : text[i] || (noBreakSpaceFlag ? noBreakSpace : '')
   ),
   array2String,
 )(mask)
 
-export const descramble = (result = "", text = "", mask = []) => R.pipe(
+export const descramble = (result = "", text = "", mask = [], noBreakSpaceFlag) => R.pipe(
   R.addIndex(R.map)((d, i) => d === 0
-    ? result[i] || ''
-    : text[i] || ''
+    ? result[i] || (noBreakSpaceFlag ? noBreakSpace : '')
+    : text[i] || (noBreakSpaceFlag ? noBreakSpace : '')
   ),
   array2String,
 )(mask)
 
-export const scramble = (result = "", _, mask = []) => R.pipe(
+export const scramble = (result = "", _, mask = [], noBreakSpaceFlag) => R.pipe(
   R.addIndex(R.map)((d, i) => d === 0
-    ? result[i] || ''
+    ? result[i] || (noBreakSpaceFlag ? noBreakSpace : '')
     : randomChar()
   ),
   array2String,